fix(sonar,entrypoint)
