+++ src/main/java/org/georchestra/urbanisme/GeorchestraLogBackConfigLoader.java
@@ -1,23 +1,21 @@
 package org.georchestra.urbanisme;
 
 import java.io.File;
-import java.io.IOException;
 
-import ch.qos.logback.classic.LoggerContext;
-import ch.qos.logback.classic.joran.JoranConfigurator;
-import ch.qos.logback.core.joran.spi.JoranException;
 import javax.annotation.PostConstruct;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
+import ch.qos.logback.classic.LoggerContext;
+import ch.qos.logback.classic.joran.JoranConfigurator;
+import ch.qos.logback.core.joran.spi.JoranException;
+
 /**
  * Simple Utility class for loading an external config file for logback
  *
- * inspired from:
- * * cadastrapp source code
- * * https://bowerstudios.com/node/896
+ * inspired from: * cadastrapp source code * https://bowerstudios.com/node/896
  */
 @Component
 public class GeorchestraLogBackConfigLoader {
@@ -29,7 +27,7 @@ public class GeorchestraLogBackConfigLoader {
 	}
 
 	@PostConstruct
-	public void init() throws IOException, JoranException {
+	public void init() throws JoranException {
 		LoggerContext lc = (LoggerContext) LoggerFactory.getILoggerFactory();
 		String georDatadir = System.getProperty("georchestra.datadir");
 		if (georDatadir == null) {
@@ -45,10 +43,12 @@ public class GeorchestraLogBackConfigLoader {
 		}
 
 		if (!urbaLogBackFile.isFile()) {
-			logger.error("Logback External Config File Parameter exists, but does not reference a file: {}", urbaLogBackFile);
+			logger.error("Logback External Config File Parameter exists, but does not reference a file: {}",
+					urbaLogBackFile);
 		} else {
 			if (!urbaLogBackFile.canRead()) {
-				logger.error("Logback External Config File exists and is a file, but cannot be read: {}", urbaLogBackFile);
+				logger.error("Logback External Config File exists and is a file, but cannot be read: {}",
+						urbaLogBackFile);
 			} else {
 				JoranConfigurator configurator = new JoranConfigurator();
 				configurator.setContext(lc);
