+++ src/main/java/org/georchestra/urbanisme/renseignurba/RenseignUrbaBackend.java
@@ -17,7 +17,7 @@
  * geOrchestra.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.georchestra.urbanisme.RenseignUrba;
+package org.georchestra.urbanisme.renseignurba;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
