+++ src/test/java/org/georchestra/urbanisme/RenseignUrba/TemplateRuleHelperTest.java
@@ -9,6 +9,7 @@ import static org.mockito.Mockito.when;
 import java.util.List;
 import java.util.Map;
 
+import org.georchestra.urbanisme.renseignurba.TemplateRuleHelper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
